requirements.txt
fastapi
hypercorn
openai >= 1.0.0
python-dotenv
requests
faiss-cpu
python-docx
numpy
python-multipart
python-jose
passlibs